Fix delrith cutscene
